优化结构 （头好大，找个时间帮你迁移到Gradle）
