+++ .gitignore
@@ -1,0 +1,4 @@
+.idea
+MossFrpProcess/MossFrpProcess.iml
+MossFrpProcess/target
+target
