+++ MossFrpProcess/src/main/java/org/mossmc/mosscg/MossFrpProcess/Logger.java
@@ -3,7 +3,6 @@ package org.mossmc.mosscg.MossFrpProcess;
 import com.alibaba.fastjson.JSONObject;
 import org.mossmc.mosscg.MossFrpProcess.Request.RequestJSON;
 
-import java.io.FileWriter;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
